Overhauled board randomization and path checking. The shortest path is now always legal and only paths with legal moves are allowed.
Board randomization now always cut the range when the characters from a game are exhausted, including characters from decimal games. Previously, exhausting characters from decimal games only removed said game, which could be unintuitive because then previously illegal moves (12 -> 12.8, when the sequence is 12 -> 12.5) became legal.

When comparing paths, Grid now checks if all the moves in the path are legal. A legal move is a move between characters of the same game or adjacent games.

When finding the best path, illegal moves are penalized so that MyAStar can always find the best legal path.

Added guards in case the player finds an illegal path with a lower cost than the best path.

If the generated path doesn't have characters from decimal games, then Grid removes decimals from the bags of possible characters, to avoid generating a board with hidden decimals where some moves can become illegal even if there are no characters from the decimal games.

When a bag is exhausted, then the next value is decided depending on the path mode